test-requirements.txt
-r requirements.txt
pytest
pytest-django
